# Udacity : Practice Questions 39 ---> my try (worked) # NANODEGREE PROGRAM - AI Programming with Python
Two Tasks:
1A: Create dictionary with the count of Oscar nominations for each director 
1B: Create dictionary with the count of Oscar wins for each director
Info:
The questions are based on data on Oscar Award Nominations for Best Director between the years 1931 to 2010. To start you off, we've provided a dictionary called "nominated" with the year (as key) and list of directors who were nominated in that year (as value). We've provided you with a different dictionary called "winners" with the year (as key) and list of directors who won the award in that year (as value).